requirements.txt
bitarray==2.9.2
blinker==1.8.2
camelot-py==0.11.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
et-xmlfile==1.1.0
Flask==3.0.3
googletrans==3.0.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.1
pdfminer.six==20231228
pdfplumber==0.11.3
pdfreader==0.1.15
pillow==10.4.0
pycparser==2.22
pycryptodome==3.20.0
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.8.2
python-pptx==1.0.2
pytz==2024.1
pywin32==306
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
XlsxWriter==3.2.0
